+++ Source_Files/GameWorld/monsters.h
@@ -291,7 +291,7 @@ struct monster_definition *get_monster_definition_external(const short type);
 // and used a growable list for the indices
 #define LOCAL_INTERSECTING_MONSTER_BUFFER_SIZE (get_dynamic_limit(_dynamic_limit_local_collision))
 #define GLOBAL_INTERSECTING_MONSTER_BUFFER_SIZE (get_dynamic_limit(_dynamic_limit_global_collision))
-bool possible_intersecting_monsters(vector<short> *IntersectedObjectsPtr, short maximum_object_count, short polygon_index, bool include_scenery);
+bool possible_intersecting_monsters(vector<short> *IntersectedObjectsPtr, unsigned maximum_object_count, short polygon_index, bool include_scenery);
 #define monsters_nearby(polygon_index) possible_intersecting_monsters(0, 0, (polygon_index), false)
 
 void get_monster_dimensions(short monster_index, world_distance *radius, world_distance *height);
